requirements.txt
networkx
nx_arangodb
python-arango
pandas
numpy
matplotlib
langgraph
langchain-openai
langchain-community
langchain-core